+++ src/expr/impl/src/lib.rs
@@ -37,6 +37,7 @@ mod aggregate;
 mod scalar;
 mod table_function;
 mod udf;
+mod window_function;
 
 /// Enable functions in this crate.
 #[macro_export]
